+++ packages/web3-types/src/index.ts
@@ -31,4 +31,3 @@ export * from './web3_api_types.js';
 export * from './web3_base_provider.js';
 export * from './web3_base_wallet.js';
 export * from './web3_deferred_promise_type.js';
-export * from './eip6963_types.js';