+++ packages/web3/src/web3_eip6963.ts
@@ -15,17 +15,26 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import type {
-	EIP6963ProviderDetail,
-	EIP6963ProvidersMapUpdateEvent,
-	EIP6963ProviderResponse,
-	Web3APISpec,
-} from 'web3-types';
+import { Web3APISpec, EIP1193Provider } from 'web3-types';
 
 export enum Eip6963EventName {
 	eip6963announceProvider = 'eip6963:announceProvider',
 	eip6963requestProvider = 'eip6963:requestProvider',
 }
+
+export interface EIP6963ProviderInfo {
+	uuid: string;
+	name: string;
+	icon: string;
+	rdns: string;
+}
+
+export interface EIP6963ProviderDetail<API = Web3APISpec> {
+	info: EIP6963ProviderInfo;
+	provider: EIP1193Provider<API>;
+}
+
+export type EIP6963ProviderResponse = Map<string, EIP6963ProviderDetail>;
 
 export interface EIP6963AnnounceProviderEvent<API = Web3APISpec> extends CustomEvent {
 	type: Eip6963EventName.eip6963announceProvider;
@@ -39,6 +48,10 @@ export interface EIP6963RequestProviderEvent extends Event {
 export const eip6963ProvidersMap: EIP6963ProviderResponse = new Map();
 
 export const web3ProvidersMapUpdated = 'web3:providersMapUpdated';
+export interface EIP6963ProvidersMapUpdateEvent extends CustomEvent {
+	type: string;
+	detail: EIP6963ProviderResponse;
+}
 
 export const requestEIP6963Providers = async (): Promise<EIP6963ProviderResponse> =>
 	new Promise((resolve, reject) => {