+++ packages/web3/src/index.ts
@@ -327,6 +327,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 import Web3 from './web3.js';
 
 export * from './types.js';
+export * from './web3_eip6963.js';
 export default Web3;
 
 /**