+++ packages/web3-core/test/unit/web3_config.test.ts
@@ -50,6 +50,7 @@ const defaultConfig = {
 	transactionBuilder: undefined,
 	transactionTypeParser: undefined,
 	customTransactionSchema: undefined,
+	ignoreGasPricing: false,
 };
 const setValue = {
 	string: 'newValue',