+++ packages/web3/test/integration/web3RPCProviders.test.ts
@@ -15,12 +15,13 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { mainnet, Network, QuickNodeProvider, Transport } from 'web3-rpc-providers';
+import { mainnet, Network, QuickNodeProvider, InfuraProvider, Transport } from 'web3-rpc-providers';
 import { Web3 } from '../../src/index';
 
+jest.setTimeout(10000);
 describe('Web3 RPC Provider Integration tests', () => {
 	const transports = Object.values(Transport);
-	const networks = [
+	const quickNodeNetworks = [
 		Network.ETH_MAINNET,
 		Network.ETH_HOLESKY,
 		Network.ETH_SEPOLIA,
@@ -33,7 +34,7 @@ describe('Web3 RPC Provider Integration tests', () => {
 	];
 
 	transports.forEach(transport => {
-		networks.forEach(network => {
+		quickNodeNetworks.forEach(network => {
 			it(`QuickNodeProvider should work with ${transport} transport and ${network} network`, async () => {
 				const provider = new QuickNodeProvider(network, transport);
 				const web3 = new Web3(provider);
@@ -46,6 +47,86 @@ describe('Web3 RPC Provider Integration tests', () => {
 					web3.provider?.disconnect();
 				}
 			});
+		});
+	});
+
+	const infuraNetworks = [
+		Network.PALM_MAINNET,
+		Network.PALM_TESTNET,
+		Network.BLAST_MAINNET,
+		Network.BLAST_SEPOLIA,
+		Network.AVALANCHE_MAINNET,
+		Network.AVALANCHE_FUJI,
+		Network.STARKNET_MAINNET,
+		Network.STARKNET_SEPOLIA,
+		Network.ZKSYNC_MAINNET,
+		Network.ZKSYNC_SEPOLIA,
+		Network.CELO_MAINNET,
+		Network.CELO_ALFAJORES,
+		Network.BSC_MAINNET,
+		Network.BSC_TESTNET,
+		Network.MANTLE_MAINNET,
+		Network.MANTLE_SEPOLIA,
+		Network.ETH_MAINNET,
+		Network.ETH_HOLESKY,
+		Network.ETH_SEPOLIA,
+		Network.ARBITRUM_MAINNET,
+		Network.ARBITRUM_SEPOLIA,
+		Network.BASE_MAINNET,
+		Network.BASE_SEPOLIA,
+		Network.BNB_MAINNET,
+		Network.BNB_TESTNET,
+		Network.LINEA_MAINNET,
+		Network.LINEA_SEPOLIA,
+		Network.POLYGON_MAINNET,
+		Network.POLYGON_AMONY,
+		Network.OPTIMISM_MAINNET,
+		Network.OPTIMISM_SEPOLIA,
+	];
+	transports.forEach(transport => {
+		infuraNetworks.forEach(network => {
+			// skip not exists endpoints
+			if (
+				!(
+					[
+						Network.PALM_MAINNET,
+						Network.PALM_TESTNET,
+						Network.BLAST_SEPOLIA,
+						Network.STARKNET_MAINNET,
+						Network.STARKNET_SEPOLIA,
+						Network.ZKSYNC_SEPOLIA,
+						Network.BSC_TESTNET,
+						Network.MANTLE_SEPOLIA,
+						Network.BNB_TESTNET,
+					].includes(network) && transport === Transport.WebSocket
+				)
+			) {
+				it.skip(`InfuraProvider should work with ${transport} transport and ${network} network`, async () => {
+					const provider = new InfuraProvider(
+						network,
+						transport,
+						process.env.INFURA_PROVIDER_KEY,
+					);
+
+					const web3 = new Web3(provider);
+					const result =
+						network === Network.STARKNET_MAINNET || network === Network.STARKNET_SEPOLIA
+							? BigInt(
+									await web3.requestManager.send({
+										method: 'starknet_blockNumber',
+										params: [],
+									}),
+							  )
+							: await web3.eth.getBlockNumber();
+
+					expect(typeof result).toBe('bigint');
+					expect(result > 0).toBe(true);
+
+					if (transport === Transport.WebSocket) {
+						web3.provider?.disconnect();
+					}
+				});
+			}
 		});
 	});
 	it(`should work with mainnet provider`, async () => {