+++ CHANGELOG.md
@@ -2728,3 +2728,9 @@ If there are any bugs, improvements, optimizations or any new feature proposal f
 -   The callback function provided to the static `Web3.onNewProviderDiscovered` function expects a parameter of type `EIP6963ProvidersMapUpdateEvent` as opposed to `EIP6963AnnounceProviderEvent`. (#7242)
 
 ## [Unreleased]
+
+### Changed
+
+#### web3-eth
+
+-   Allow `getEthereumjsTxDataFrom` to return additional fields that may be passed if using a `customTransactionSchema`.