+++ CHANGELOG.md
@@ -2731,9 +2731,17 @@ If there are any bugs, improvements, optimizations or any new feature proposal f
 
 ### Added
 
+#### web3-core
+
+-   Added new property `ignoreGasPricing` to `Web3ConfigOptions`. If `ignoreGasPricing` is true, gasPrice will not be estimated (#7320)
+
 #### web3-eth
 
--   `syncing` subscription now supports Besu and Nethermind payload format
+-   `populateGasPrice` function now checks `Web3Context.config.ignoreGasPricing`. If `ignoreGasPricing` is true, gasPrice will not be estimated (#7320)
+
+#### web3-rpc-providers
+
+-   InfuraProvider was added (#7286)
 
 ### Changed
 
@@ -2741,6 +2749,20 @@ If there are any bugs, improvements, optimizations or any new feature proposal f
 
 -   Allow `getEthereumjsTxDataFrom` to return additional fields that may be passed if using a `customTransactionSchema`.
 
+#### web3-types
+
+-   update the type for `baseFeePerGas` at `web3.eth.getFeeHistory` to be a number. (#7291)
+
+### Fixed
+
 #### web3
 
 -   Exported EIP-6963 types are no longer under the `providers` namespace. (#7270)
+
+#### web3-eth-abi
+
+-   `decodeLog` , `decodeParametersWith` , `decodeParameters` and `decodeParameters` now accepts first immutable param as well (#7288)
+
+#### web3-utils
+
+-   fix `padRight` validation failure on large `uint` (#7265)